requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.7
sqlparse==0.4.3
